change data dtype in evaluation to np.uint8, reduce time and memory usage

+++ evaluation/source/dataset.py
@@ -76,7 +76,7 @@ class Dataset(object):
 
     def _get_all_elements(self, sequence, obj_type):
         obj = np.array(Image.open(self.sequences[sequence][obj_type][0]))
-        all_objs = np.zeros((len(self.sequences[sequence][obj_type]), *obj.shape))
+        all_objs = np.zeros((len(self.sequences[sequence][obj_type]), *obj.shape), dtype=np.uint8)
         obj_id = []
         for i, obj in enumerate(self.sequences[sequence][obj_type]):
             all_objs[i, ...] = np.array(Image.open(obj))
@@ -97,8 +97,8 @@ class Dataset(object):
 
         if separate_objects_masks:
             num_objects = int(np.max(masks[0, ...]))
-            tmp = np.ones((num_objects, *masks.shape))
-            tmp = tmp * np.arange(1, num_objects + 1)[:, None, None, None]
+            tmp = np.ones((num_objects, *masks.shape), dtype=masks.dtype)
+            tmp = tmp * np.arange(1, num_objects + 1, dtype=tmp.dtype)[:, None, None, None]
             masks = (tmp == masks[None, ...])
             masks = masks > 0
         return masks, masks_void, masks_id
